Alt+arrows in tmux moves to the pane in that direction
- Alt+arrow is vim-agnostic (i.e. it will only switch tmux panes).
- This is added for a specific use case: you are running vim on a remote server
with ssh. In this case, Alt+hjkl will cause vim to try to move to the next tmux
pane *on the remote server* (using a 'system' call), which is obviously not
desirable. As a workaround Alt+arrows will force tmux pane movement on the
local machine. Note that on my keyboard, CapsLock+hjkl are arrow keys,
making the woraround physically easy to use.